fix: Odometrie-Reset vor CMDP-Wegpruefung absichern

+++ 5_System/src/Printer.h
@@ -19,8 +19,8 @@ public:
     void printCompletedFrame(
         VehicleController& vehicle,
         const RearPendingFrame& frame,
-        int16_t hiLi_i_cms,
-        int16_t hiRe_i_cms,
+        float hiLi_i_cms,
+        float hiRe_i_cms,
         int16_t hiLi_pwm,
         int16_t hiRe_pwm);
 
@@ -30,37 +30,43 @@ public:
         const MecanumOdometer& odom);
 
 #ifdef PRINTER_MODE_CHASSIS
-    void printWheels(VehicleController& vehicle,
-        float v2_ist_cms, float v3_ist_cms,
+    void printWheels(
+        VehicleController& vehicle,
+        float v2_ist_cms,
+        float v3_ist_cms,
         uint32_t t_ms);
 
-    void printFrame(VehicleController& vehicle,
+    void printFrame(
+        VehicleController& vehicle,
         uint32_t t_ms,
-        int16_t voLi_i_cms,
-        int16_t voRe_i_cms,
-        int16_t hiLi_i_cms,
-        int16_t hiRe_i_cms);
+        float voLi_i_cms,
+        float voRe_i_cms,
+        float hiLi_i_cms,
+        float hiRe_i_cms);
 #endif
 
 #ifdef PRINTER_MODE_RAEDER
-    void printWheels(VehicleController& vehicle,
-        float v2_ist_cms, float v3_ist_cms,
-        int16_t pwm2, int16_t pwm3,
+    void printWheels(
+        VehicleController& vehicle,
+        float v2_ist_cms,
+        float v3_ist_cms,
+        int16_t pwm2,
+        int16_t pwm3,
         uint32_t t_ms);
 
     void printFrame(
         uint32_t t_ms,
-        int16_t voLi_s_cms,
-        int16_t voLi_i_cms,
+        float voLi_s_cms,
+        float voLi_i_cms,
         int16_t voLi_pwm,
-        int16_t voRe_s_cms,
-        int16_t voRe_i_cms,
+        float voRe_s_cms,
+        float voRe_i_cms,
         int16_t voRe_pwm,
-        int16_t hiLi_s_cms,
-        int16_t hiLi_i_cms,
+        float hiLi_s_cms,
+        float hiLi_i_cms,
         int16_t hiLi_pwm,
-        int16_t hiRe_s_cms,
-        int16_t hiRe_i_cms,
+        float hiRe_s_cms,
+        float hiRe_i_cms,
         int16_t hiRe_pwm);
 #endif
 };
